security(mcp): bind providers on label boundaries, clamp tool params, lock the credential store
Provider binding used ends_with("openai.com"), so evil-openai.com,
xgithub.com and any host ending in the provider string were treated as the
provider: a brokered reference sent to such a host was dereferenced into the
real key. host_is_or_under matches on a label boundary, case-insensitively,
with the FQDN trailing dot stripped, and is the one place a destination is
compared against a provider domain.

The builtin tools read start_index, max_length, context_lines and
max_matches as u64 and cast them to usize; max_length: 18446744073709551615
overflowed the slice arithmetic and panicked the builtin subprocess. Every
such parameter is clamped through bounded_param and the offset arithmetic is
saturating. snapshots_history sized files with a plain metadata call that
followed symlinks; it now walks through ContainedDir like the other snapshot
tools and refuses symlinks and special files.

The durable credential store serialized its load-insert-write with a
process-local mutex. Every capsem-process on the host captures into the same
file, so two processes interleaved and the later rename dropped everything
the earlier one added. The store now takes flock(2) on a sibling lock file
for each read-modify-write.

capsem-mcp logged ?params for create and run, and the derived Debug printed
every environment value, API keys included, into ~/.capsem/run/mcp.log. The
parameter types move to params.rs with a Debug that names keys only.

file_tools.rs splits its snapshot handlers into file_tools/snapshots.rs and
the test modules move their fixture-backed cases into submodules, all pure
moves under the line-count ratchet.
