+++ config/gate.toml
@@ -1069,9 +1069,6 @@ must_stay_below_lines = 3000
 "crates/capsem-core/src/hypervisor/kvm/virtio_blk.rs" = 1856
 "crates/capsem-core/src/hypervisor/kvm/virtio_blk/tests.rs" = 1059
 "crates/capsem-core/src/hypervisor/kvm/virtio_fs/tests.rs" = 1997
-"crates/capsem-core/src/mcp/builtin_tools/tests.rs" = 1678
-"crates/capsem-core/src/mcp/file_tools.rs" = 1172
-"crates/capsem-core/src/mcp/file_tools/tests.rs" = 1092
 "crates/capsem-core/src/net/mitm_proxy/mod.rs" = 2441
 "crates/capsem-core/src/net/mitm_proxy/telemetry_hook/tests.rs" = 1188
 "crates/capsem-core/src/net/policy_config/profile_contract.rs" = 2163
@@ -1087,8 +1084,6 @@ must_stay_below_lines = 3000
 "crates/capsem-logger/src/session_index/tests.rs" = 1205
 "crates/capsem-logger/src/writer.rs" = 1496
 "crates/capsem-logger/src/writer/tests.rs" = 2002
-"crates/capsem-mcp/src/main.rs" = 1033
-"crates/capsem-mcp/src/tests.rs" = 1019
 "crates/capsem-mock-server/src/main.rs" = 1611
 "crates/capsem-process/src/vsock.rs" = 1690
 "crates/capsem-proto/src/tests.rs" = 1591
