+++ crates/capsem-credentials/src/durable.rs
@@ -1,6 +1,6 @@
 use std::collections::HashMap;
+use std::fs::File;
 use std::path::{Path, PathBuf};
-use std::sync::{Mutex, OnceLock};
 
 use tracing::warn;
 
@@ -8,8 +8,6 @@ use crate::provider::credential_provider_from_str;
 use crate::{credential_store_account, CredentialProvider};
 
 pub const STORE_PATH_ENV: &str = "CAPSEM_CREDENTIAL_STORE_PATH";
-
-static STORE_LOCK: OnceLock<Mutex<()>> = OnceLock::new();
 
 #[cfg(test)]
 mod tests;
@@ -60,26 +58,19 @@ fn disk_store_write(
     credential_ref: &str,
     raw_value: &str,
 ) -> Result<(), String> {
-    let _guard = store_lock()
-        .lock()
-        .map_err(|_| "credential disk store lock poisoned".to_string())?;
+    let _guard = StoreLock::exclusive(path)?;
     let mut map = disk_store_load(path)?;
     map.insert(
         credential_store_account(provider, credential_ref),
         raw_value.to_string(),
     );
-    if let Some(parent) = path.parent() {
-        std::fs::create_dir_all(parent).map_err(|error| format!("create credential store dir: {error}"))?;
-    }
     let json =
         serde_json::to_string_pretty(&map).map_err(|error| format!("serialize credential disk store: {error}"))?;
     write_secret_file(path, json.as_bytes())
 }
 
 fn disk_store_read(path: &Path, provider: CredentialProvider, credential_ref: &str) -> Result<String, String> {
-    let _guard = store_lock()
-        .lock()
-        .map_err(|_| "credential disk store lock poisoned".to_string())?;
+    let _guard = StoreLock::shared(path)?;
     let map = disk_store_load(path)?;
     let account = credential_store_account(provider, credential_ref);
     map.get(&account)
@@ -88,9 +79,7 @@ fn disk_store_read(path: &Path, provider: CredentialProvider, credential_ref: &s
 }
 
 fn disk_store_hydrate(path: &Path) -> Result<Vec<(CredentialProvider, String, String)>, String> {
-    let _guard = store_lock()
-        .lock()
-        .map_err(|_| "credential disk store lock poisoned".to_string())?;
+    let _guard = StoreLock::shared(path)?;
     let map = disk_store_load(path)?;
     let mut entries = Vec::new();
     for (account, raw_value) in map {
@@ -103,8 +92,53 @@ fn disk_store_hydrate(path: &Path) -> Result<Vec<(CredentialProvider, String, St
     Ok(entries)
 }
 
-fn store_lock() -> &'static Mutex<()> {
-    STORE_LOCK.get_or_init(|| Mutex::new(()))
+/// Cross-process lock on the store, held for one read-modify-write.
+///
+/// Every `capsem-process` on the host captures into the same file, and a
+/// capture is load -> insert -> write-temp -> rename. A process-local mutex
+/// orders that inside one process only; two processes interleave and the
+/// later rename drops everything the earlier one added. The lock is an
+/// `flock(2)` on a sibling file (never the store itself: the store is renamed
+/// into place, so a lock on its inode would not survive the swap) and is
+/// taken before the load, so the map written is the map that was read.
+struct StoreLock {
+    _file: File,
+}
+
+impl StoreLock {
+    fn exclusive(store: &Path) -> Result<Self, String> {
+        let file = Self::open(store)?;
+        file.lock().map_err(|error| format!("lock credential store: {error}"))?;
+        Ok(Self { _file: file })
+    }
+
+    fn shared(store: &Path) -> Result<Self, String> {
+        let file = Self::open(store)?;
+        file.lock_shared()
+            .map_err(|error| format!("lock credential store for reading: {error}"))?;
+        Ok(Self { _file: file })
+    }
+
+    fn open(store: &Path) -> Result<File, String> {
+        let parent = store
+            .parent()
+            .ok_or_else(|| "credential store path has no parent directory".to_string())?;
+        std::fs::create_dir_all(parent).map_err(|error| format!("create credential store dir: {error}"))?;
+        let file_name = store
+            .file_name()
+            .and_then(|name| name.to_str())
+            .unwrap_or("credential-store.json");
+        let mut options = std::fs::OpenOptions::new();
+        options.read(true).write(true).create(true).truncate(false);
+        #[cfg(unix)]
+        {
+            use std::os::unix::fs::OpenOptionsExt;
+            options.mode(0o600);
+        }
+        options
+            .open(parent.join(format!(".{file_name}.lock")))
+            .map_err(|error| format!("open credential store lock: {error}"))
+    }
 }
 
 fn disk_store_load(path: &Path) -> Result<HashMap<String, String>, String> {
